Reverting back to specific container component for Table
