+++ src/components/Table/presenter.js
@@ -1,0 +1,20 @@
+import React from 'react';
+import Header from '../Header';
+import Item from '../Item';
+
+const Table = ({result}) => {
+  return (
+    <div className="table">
+      <Header/>
+      {
+        result.hits.map((item, key) => {
+          return (
+            <Item key={key} item={item}/>
+          );
+        })
+      }
+    </div>
+  );
+};
+
+export default Table;
